split misleadingly named test file
